[SDKUNITY-4217] Expose raw audio buffers via WebRTC (#54)
- Added delegates to MLWebRTC.AudioSink to provide the app with audio
buffers when the BufferNotifyMode is either "NotifyAndPlay" or
"NotifyOnly".
- When selecting "NotifyAndPlay", the app will receive
buffers from WebRTC but WebRTC will still play the audio. This is
useful for cases where apps require access to the audio buffers to
calculate something like the RMS but not take on the responsibility
of rendering the audio.
- When selecting "NotifyOnly", its the apps responsibility to render
the received audio buffers.